+++ src/Fable.Cli/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * [JS/TS] Fixed TimeSpan.FromMilliseconds (#3815) (by @ncave)
 * [Python] Fixed quotation for union string cases (by @dbrattli)
+* [Python] Fixed casing issues with identifiers and reflection info (#3811) (by @dbrattli)
 
 ## 4.17.0 - 2024-04-23